style: move content from old crate into new crate
