+++ Cargo.toml
@@ -24,15 +24,10 @@ category = "Developer Tool"
 long_description = "OxidUi is a framework for building terminal user interfaces (TUIs) in Rust. It provides a set of tools and components to create interactive and visually appealing TUIs, making it easier for developers to build command-line applications with rich user interfaces."
 
 [workspace]
-members = [
-    "crates/oxidui_style",
-    "crates/oxidui_macros",
-    "crates/termoxide_layout",
-]
+members = ["crates/oxidui_macros", "crates/termoxide_layout"]
 
 [dependencies]
 oxidui_macros = { path = "crates/oxidui_macros" }
-oxidui_style = { path = "crates/oxidui_style", default-features = true }
 
 # Path handling
 walkdir = "2.5"
@@ -47,7 +42,7 @@ enum_cycling = "0.2.1"
 
 [features]
 default = ["ratatui"]
-ratatui = ["oxidui_style/ratatui"]
+ratatui = []
 
 [profile.release]
 opt-level = 3
