--- crates/oxidui_style/Cargo.toml
@@ -1,11 +1,0 @@
-[package]
-name = "oxidui_style"
-version = "0.1.0"
-edition = "2024"
-
-[dependencies]
-ratatui = { version = "0.26", optional = true, features = ["crossterm"] }
-
-[features]
-default = ["ratatui"]
-ratatui = ["dep:ratatui"]
