+++ crates/termoxide_layout/Cargo.toml
@@ -4,5 +4,9 @@ version = "0.1.0"
 edition = "2024"
 
 [dependencies]
+ratatui = { version = "0.26", optional = true, features = ["crossterm"] }
 taffy = "0.9.2"
-oxidui_style = { path = "../oxidui_style" }
+
+[features]
+default = ["ratatui"]
+ratatui = ["dep:ratatui"]
