+++ crates/termoxide_layout/src/unit.rs
@@ -27,6 +27,7 @@
 /// # Examples
 ///
 /// ```rust
+/// use termoxide_layout::unit::Unit;
 /// let w    = Unit::cells(40);    // exactly 40 columns
 /// let h    = Unit::percent(50);  // 50% of parent height
 /// let flex = Unit::fill(1);      // take 1 share of remaining space
